requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
python-dotenv
redis
ylogging
ypostgres_lib